+++ photon-model/src/main/java/com/vmware/photon/controller/model/constants/PhotonModelConstants.java
@@ -35,7 +35,8 @@ public class PhotonModelConstants {
     public static final String STATUS_CHECK_FAILED_COUNT_INSTANCE = "StatusCheckFailedCount_Instance";
     public static final String STATUS_CHECK_FAILED_COUNT_SYSTEM = "StatusCheckFailedCount_System";
     public static final String ESTIMATED_CHARGES = "EstimatedCharges";
-    public static final String BURN_RATE_PER_HOUR = "BurnRatePerHour";
+    public static final String CURRENT_BURN_RATE_PER_HOUR = "CurrentBurnRatePerHour";
+    public static final String AVERAGE_BURN_RATE_PER_HOUR = "AverageBurnRatePerHour";
 
     public static final String DISK_WRITE_TIME_SECONDS = "DiskWriteTimeSeconds";
     public static final String DISK_READ_TIME_SECONDS = "DiskReadTimeSeconds";
@@ -73,7 +74,8 @@ public class PhotonModelConstants {
         statMap.put(STATUS_CHECK_FAILED_COUNT_INSTANCE, UNIT_COUNT);
         statMap.put(STATUS_CHECK_FAILED_COUNT_SYSTEM, UNIT_COUNT);
         statMap.put(ESTIMATED_CHARGES, UNIT_COST);
-        statMap.put(BURN_RATE_PER_HOUR, UNIT_COST);
+        statMap.put(CURRENT_BURN_RATE_PER_HOUR, UNIT_COST);
+        statMap.put(AVERAGE_BURN_RATE_PER_HOUR, UNIT_COST);
         statMap.put(DISK_WRITE_TIME_SECONDS, UNIT_SECONDS);
         statMap.put(DISK_READ_TIME_SECONDS, UNIT_SECONDS);
         statMap.put(MEMORY_AVAILABLE_BYTES, UNIT_BYTES);